requirements.txt
streamlit
faiss-cpu
numpy
pandas
sentence-transformers
pdfplumber
langchain
langchain-community
langchain-ollama
ollama
